+++ CHANGELOG.md
@@ -6,3 +6,9 @@
 > ## Version 0.1.1
 
 * Fix template variables
+
+> ## Version 0.1.2
+
+* Add error handling to the API responses
+* Add Message model to make response easier to navigate
+* Add Template model to make response easier to navigate